Add feedback and subdirectory options to download all as zip
If feedback is checked, a feedback.html file for each student
will be included in the zip. It includes the grade, feedback,
date and teacher's name.

If subdirectory is checked, a folder named after student's name
will be created and all files of the students will be put in
the folder

glovenone@github made the first version of this patch.