Add feedback and subfolder to download all as zip
A feedback.html file for each student will be included in the zip. It includes the grade, feedback, date and teacher's name.

And a folder named after student's name will be created and all files of the students will be put in the folder

glovenone@github made the first version of this patch.